Move `DataflowNodeToSqlCteVisitor` to a separate file.
